requirements.txt
pyre2==0.3.6
python-magic==0.4.24
